pyproject.toml
[project]
name = "self-discover-prompt"
version = "0.1.0"
description = "Python implementation of the SELF-DISCOVER method: https://arxiv.org/abs/2402.03620"
authors = [
    { name = "satoooh", email = "bighope.lumiere0916@gmail.com" }
]
dependencies = [
    "langchain>=0.1.16",
    "langchain-openai>=0.1.3",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/"]
